+++ components/user-mgt/org.wso2.carbon.user.mgt/src/main/java/org/wso2/carbon/user/mgt/RolePermissionManagementServiceImpl.java
@@ -22,7 +22,7 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.context.CarbonContext;
 import org.wso2.carbon.user.core.UserRealm;
-import org.wso2.carbon.user.mgt.common.PermissionValue;
+import org.wso2.carbon.user.mgt.common.Permission;
 import org.wso2.carbon.user.mgt.common.UIPermissionNode;
 import org.wso2.carbon.user.mgt.common.UserAdminException;
 
@@ -62,7 +62,7 @@ public class RolePermissionManagementServiceImpl implements RolePermissionManage
     }
 
     @Override
-    public PermissionValue[] getAllPermissions(int tenantId) throws RolePermissionException {
+    public Permission[] getAllPermissions(int tenantId) throws RolePermissionException {
 
         try {
             return getAllPermissions(getUserAdminProxy().getAllUIPermissions(tenantId));
@@ -114,21 +114,21 @@ public class RolePermissionManagementServiceImpl implements RolePermissionManage
      * Recursively go through UIPermissionNode.
      *
      * @param node UIPermissionNode of permissions.
-     * @return  String[] of permissions.
+     * @return  Permission[] of permissions.
      */
-    private PermissionValue[] getAllPermissions(UIPermissionNode node) {
+    private Permission[] getAllPermissions(UIPermissionNode node) {
 
-        List<PermissionValue> permissions = new ArrayList<>();
+        List<Permission> permissions = new ArrayList<>();
         UIPermissionNode[] childNodes = node.getNodeList();
-        PermissionValue permissionValue = new PermissionValue();
-        permissionValue.setDisplayName(node.getDisplayName());
-        permissionValue.setResourcePath(node.getResourcePath());
-        permissions.add(permissionValue);
+        Permission permission = new Permission();
+        permission.setDisplayName(node.getDisplayName());
+        permission.setResourcePath(node.getResourcePath());
+        permissions.add(permission);
         if (ArrayUtils.isNotEmpty(childNodes)) {
             for (UIPermissionNode childNode : childNodes) {
                 permissions.addAll(Arrays.asList(getAllPermissions(childNode)));
             }
         }
-        return permissions.toArray(new PermissionValue[0]);
+        return permissions.toArray(new Permission[0]);
     }
 }
